requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.7
Django==3.2.18
django-crispy-forms==2.0
django-currentuser==0.5.3
django-filter==23.1
django-phonenumber-field==7.0.2
kombu==5.2.4
phonenumberslite==8.13.10
Pillow==9.5.0
prompt-toolkit==3.0.38
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
vine==5.0.0
wcwidth==0.2.6
